Closes #5654 Update WebP detection for browsers missing Accept headers data (#5664)
* move webp subscriber
* add webp admin subscriber
* Refactor WebP subscriber
* fix PHPCS
* move import
* update condition
* update unit tests
* fix group name
* remove unused file
* update unit test
* remove trailingslashit
* update convertToWebP test
* fix path
* update class instance check
* fix cap
* add missing trailingslash for content dir path
* improve code
* correctly return webp filename
* code improvements
* deprecate filter
* remove unused continue
* return true early if matching user agent and version
* add early return
