+++ inc/classes/Buffer/class-cache.php
@@ -604,24 +604,79 @@ class Cache extends Abstract_Buffer {
 			return $filename;
 		}
 
-		$http_accept = $this->config->get_server_input( 'HTTP_ACCEPT', '' );
-
-		if ( ! $http_accept && function_exists( 'apache_request_headers' ) ) {
-			$headers     = apache_request_headers();
-			$http_accept = isset( $headers['Accept'] ) ? $headers['Accept'] : '';
-		}
-
-		if ( ! $http_accept || false === strpos( $http_accept, 'webp' ) ) {
-			if ( preg_match( '#Firefox/(?<version>[0-9]{2})#i', $this->config->get_server_input( 'HTTP_USER_AGENT' ), $matches ) ) {
-				if ( 66 <= (int) $matches['version'] ) {
-					return $filename . '-webp';
-				}
-			}
-
+		if ( ! $this->is_browser_webp_compatible() ) {
 			return $filename;
 		}
 
 		return $filename . '-webp';
+	}
+
+	/**
+	 * Checks if the browser is WebP compatible
+	 *
+	 * @since 3.12.6
+	 *
+	 * @return bool
+	 */
+	private function is_browser_webp_compatible(): bool {
+		// Only to supporting browsers.
+		$http_accept = $this->config->get_server_input( 'HTTP_ACCEPT', '' );
+
+		if (
+			empty( $http_accept )
+			&&
+			function_exists( 'apache_request_headers' )
+		) {
+			$headers     = apache_request_headers();
+			$http_accept = isset( $headers['Accept'] ) ? $headers['Accept'] : '';
+		}
+
+		if (
+			! empty( $http_accept )
+			&&
+			false !== strpos( $http_accept, 'webp' )
+		) {
+			return true;
+		}
+
+		return $this->is_user_agent_compatible();
+	}
+
+	/**
+	 * Check the User Agent if the Accept headers is missing the WebP info
+	 *
+	 * @since 3.12.6
+	 *
+	 * @return bool
+	 */
+	private function is_user_agent_compatible(): bool {
+		$user_agent = $this->config->get_server_input( 'HTTP_USER_AGENT' );
+
+		if ( empty( $user_agent ) ) {
+			return false;
+		}
+
+		if ( preg_match( '#Firefox/(?<version>[0-9]{2,})#i', $user_agent, $matches ) ) {
+			if ( 66 >= (int) $matches['version'] ) {
+				return false;
+			}
+		}
+
+		if ( preg_match( '#(?:iPad|iPhone)(.*)Version/(?<version>[0-9]{2,})#i', $user_agent, $matches ) ) {
+			if ( 14 > (int) $matches['version'] ) {
+				return false;
+			}
+
+			return true;
+		}
+
+		if ( preg_match( '#Version/(?<version>[0-9]{2,})(?:.*)Safari#i', $user_agent, $matches ) ) {
+			if ( 16 > (int) $matches['version'] ) {
+				return false;
+			}
+		}
+
+		return true;
 	}
 
 	/**
