+++ tests/Fixtures/inc/Addon/WebP/Subscriber/nomatch.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en-US" dir="ltr">
+<head>
+	<meta property="og:image" data-url="https://example.com/wp-content/uploads/2019/09/one-image.png"/>
+	<link rel="image_src" data-foohref="https://example.com/wp-content/uploads/2019/09/one-image.png" />
+	<link rel="apple-touch-icon" data-barhref="https://cdn-example.net/wp-content/uploads/2017/02/apple-touch-icon.png" />
+	<link rel="icon" type="image/png" data-newhref="https://example.com/wp-content/uploads/2017/02/favicon-32x32.png" sizes="32x32" />
+	<meta name="msapplication-TileImage" data-lazycontent="https://example.com/wp-content/uploads/2017/02/mstile-144x144.png" />
+</head>
+<body>
+	<img width="200" height="200" data-ssrc="data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7" data-lazysrc="https://example.com/wp-content/uploads/2019/09/one-image.png" class="attachment-thumbnail portrait wp-post-image" alt="" itemprop="contentUrl" data-lazysrcset="https://example.com/wp-content/uploads/2019/09/one-image.png 200w, https://example.com/wp-content/uploads/2019/09/one-image-60x60.png 60w" data-lazysizes="(max-width: 200px) 100vw, 200px" />
+	<noscript>
+		<img width="200" height="200" data-ssrc="https://example.com/wp-content/uploads/2019/09/one-image.png" class="attachment-thumbnail portrait wp-post-image" alt="" itemprop="contentUrl" data-ssrcset="https://example.com/wp-content/uploads/2019/09/one-image.png 200w, https://example.com/wp-content/uploads/2019/09/one-image-60x60.png 60w" sizes="(max-width: 200px) 100vw, 200px" />
+	</noscript>
+	<img data-msrc="https://example.com/wp-content/uploads/2017/02/stats-php.gif?idsite=1" alt="" />
+</body>
+</html>
