+++ tests/Fixtures/inc/Addon/WebP/Subscriber/syncWebpCacheWithThirdPartyPlugins.php
@@ -1,0 +1,19 @@
+<?php
+
+return [
+	'shouldNotDisableWebpCacheOptionWhenWebPNotEnabled' => [
+		'option' => false,
+		'serving' => true,
+		'expected' => false,
+	],
+	'shouldNotDisableWebpCacheOptionWhenNoPluginServingWebP' => [
+		'option' => true,
+		'serving' => false,
+		'expected' => false,
+	],
+	'shouldDisableWebpCacheOptionWhenWebPEnabledAndPluginServingWebP' => [
+		'option' => true,
+		'serving' => true,
+		'expected' => true,
+	],
+];
